Split textStyle.cpp
