+++ core/src/style/textStyle.cpp
@@ -5,54 +5,32 @@
 #include "tile/tile.h"
 #include "gl/shaderProgram.h"
 #include "gl/vboMesh.h"
+#include "gl/texture.h"
+#include "gl/renderState.h"
 #include "view/view.h"
 #include "data/propertyItem.h" // Include wherever Properties is used!
 #include "labels/labelMesh.h"
-#include "gl/texture.h"
-#include "gl/renderState.h"
+#include "labels/labelContainer.h"
+#include "labels/textLabel.h"
+#include "text/fontContext.h"
+
+#include <mutex>
 
 #include "platform.h"
 #include "tangram.h"
 
-#include "alfons/alfons.h"
-#include "alfons/fontManager.h"
-#include "alfons/atlas.h"
-#include "alfons/textBatch.h"
-#include "alfons/textShaper.h"
-#include "alfons/font.h"
-#include "alfons/inputSource.h"
-
-#define SDF_IMPLEMENTATION
-#include "sdf.h"
-
-#include <bitset>
-#include <mutex>
-
-namespace alf = alfons;
-
 namespace Tangram {
 
-constexpr int texture_size = 256;
-constexpr int max_textures = 64;
+TextStyle::TextStyle(std::string _name, bool _sdf, Blending _blendMode, GLenum _drawMode) :
+    Style(_name, _blendMode, _drawMode), m_sdf(_sdf),
+    m_context(std::make_shared<AlfonsContext>())
+{
+}
 
-// #define DEFAULT_FONT "fonts/amiri-400regular.ttf"
-// #define DEFAULT_FONT "/usr/share/fonts/TTF/DejaVuSans.ttf"
-// #define DEFAULT_FONT "fonts/open sans-300normal.ttf"
-// #define FALLBACK_FONT "fonts/roboto-regular.ttf"
+TextStyle::~TextStyle() {}
 
-#define DEFAULT "fonts/NotoSans-Regular.ttf"
-#define FONT_AR "fonts/NotoNaskh-Regular.ttf"
-#define FONT_HE "fonts/NotoSansHebrew-Regular.ttf"
-#define FONT_JA "fonts/DroidSansJapanese.ttf"
-#define FALLBACK "fonts/DroidSansFallback.ttf"
-
-#define FONT_SIZE 24
-#define SDF_WIDTH 3
-
-#define ANDROID_FONT_PATH "/system/fonts/"
-
-auto vertexLayout() {
-    return std::shared_ptr<VertexLayout>(new VertexLayout({
+void TextStyle::constructVertexLayout() {
+    m_vertexLayout = std::shared_ptr<VertexLayout>(new VertexLayout({
         {"a_position", 2, GL_SHORT, false, 0},
         {"a_uv", 2, GL_UNSIGNED_SHORT, false, 0},
         {"a_color", 4, GL_UNSIGNED_BYTE, true, 0},
@@ -61,273 +39,8 @@ auto vertexLayout() {
         {"a_alpha", 1, GL_SHORT, true, 0},
         {"a_rotation", 1, GL_SHORT, false, 0},
     }));
-}
 
-struct GlyphQuad {
-    struct {
-        glm::i16vec2 pos;
-        glm::u16vec2 uv;
-    } quad[4];
-    // TODO color and stroke must not be stored per quad
-    uint32_t color;
-    uint32_t stroke;
-    alf::AtlasID atlas;
-};
-
-struct TextMesh : public LabelMesh {
-    using LabelMesh::LabelMesh;
-
-    int bufferCapacity = 0;
-    std::vector<Label::Vertex> m_vertices;
-
-    void pushQuad(GlyphQuad& _quad, Label::Vertex::State& _state) {
-
-        m_vertices.resize(m_nVertices + 4);
-
-        for (int i = 0; i < 4; i++) {
-            Label::Vertex& v = m_vertices[m_nVertices+i];
-            v.pos = _quad.quad[i].pos;
-            v.uv = _quad.quad[i].uv;
-            v.color = _quad.color;
-            v.stroke = _quad.stroke;
-            v.state = _state;
-        }
-        m_nVertices += 4;
-    }
-
-    void myUpload() {
-
-        const size_t maxVertices = 16384;
-
-        if (m_nVertices == 0) { return; }
-
-        m_vertexOffsets.clear();
-        for (size_t offset = 0; offset < m_nVertices; offset += maxVertices) {
-            size_t nVertices = maxVertices;
-            if (offset + maxVertices > m_nVertices) {
-                nVertices = m_nVertices - offset;
-            }
-            m_vertexOffsets.emplace_back(nVertices / 4 * 6, nVertices);
-        }
-
-        if (!checkValidity()) {
-            loadQuadIndices();
-            bufferCapacity = 0;
-        }
-
-        // Generate vertex buffer, if needed
-        if (m_glVertexBuffer == 0) { glGenBuffers(1, &m_glVertexBuffer); }
-
-        // Buffer vertex data
-        int vertexBytes = m_nVertices * m_vertexLayout->getStride();
-
-        RenderState::vertexBuffer(m_glVertexBuffer);
-
-        if (vertexBytes > bufferCapacity) {
-            bufferCapacity = vertexBytes;
-
-            glBufferData(GL_ARRAY_BUFFER, vertexBytes,
-                         reinterpret_cast<GLbyte*>(m_vertices.data()),
-                         m_hint);
-        } else {
-            glBufferSubData(GL_ARRAY_BUFFER, 0, vertexBytes,
-                            reinterpret_cast<GLbyte*>(m_vertices.data()));
-        }
-        m_isCompiled = true;
-        m_isUploaded = true;
-        m_dirty = false;
-    }
-
-    bool ready() { return m_isCompiled; }
-
-    void clear() {
-        m_nVertices = 0;
-        m_vertices.clear();
-        m_isCompiled = false;
-    }
-};
-
-
-struct GlyphBatch {
-
-    GlyphBatch(std::shared_ptr<VertexLayout> _vertexLayout)
-        : texture(texture_size, texture_size) {
-
-        texData.resize(texture_size * texture_size);
-        mesh = std::make_unique<TextMesh>(_vertexLayout, GL_TRIANGLES);
-    }
-
-    std::vector<unsigned char> texData;
-    Texture texture;
-    bool dirty;
-    std::unique_ptr<TextMesh> mesh;
-
-    size_t refCount = 0;
-};
-
-struct AlfonsContext : public alf::TextureCallback {
-    AlfonsContext() :
-        m_atlas(*this, texture_size) {
-
-#if defined(PLATFORM_ANDROID)
-
-        auto fontPath = systemFontPath("sans-serif", "400", "normal");
-        LOG("FONT %s", fontPath.c_str());
-
-        m_font = m_alfons.addFont(fontPath, FONT_SIZE);
-
-        std::string fallback = "";
-        int importance = 0;
-
-        while (importance < 100) {
-            fallback = systemFontFallbackPath(importance++, 400);
-            if (fallback.empty()) {
-                break;
-            }
-            if (fallback.find("UI-") != std::string::npos) {
-                continue;
-            }
-            fontPath = ANDROID_FONT_PATH;
-            fontPath += fallback;
-            LOG("FALLBACK %s", fontPath.c_str());
-
-            m_font->addFace(m_alfons.getFontFace(alf::InputSource(fontPath), FONT_SIZE));
-        }
-#else
-        m_font = m_alfons.addFont(DEFAULT, FONT_SIZE);
-        m_font->addFace(m_alfons.getFontFace(alf::InputSource(FONT_AR), FONT_SIZE));
-        m_font->addFace(m_alfons.getFontFace(alf::InputSource(FONT_HE), FONT_SIZE));
-        m_font->addFace(m_alfons.getFontFace(alf::InputSource(FONT_JA), FONT_SIZE));
-        m_font->addFace(m_alfons.getFontFace(alf::InputSource(FALLBACK), FONT_SIZE));
-#endif
-
-        m_vertexLayout = vertexLayout();
-    }
-
-    // Synchronized on m_mutex on tile-worker threads
-    void addTexture(alf::AtlasID id, uint16_t width, uint16_t height) override {
-        m_batches.emplace_back(m_vertexLayout);
-    }
-
-    // Synchronized on m_mutex, called tile-worker threads
-    void addGlyph(alf::AtlasID id, uint16_t gx, uint16_t gy, uint16_t gw, uint16_t gh,
-                  const unsigned char* src, uint16_t pad) override {
-
-        auto& texData = m_batches[id].texData;
-        auto& texture = m_batches[id].texture;
-        m_batches[id].dirty = true;
-
-        uint16_t stride = texture_size;
-        uint16_t width = texture_size;
-
-        unsigned char* dst = &texData[(gx + pad) + (gy + pad) * stride];
-
-        size_t pos = 0;
-        for (uint16_t y = 0; y < gh; y++) {
-            size_t row = (y * stride);
-            for (uint16_t x = 0; x < gw; x++) {
-                dst[row + x] = src[pos++];
-            }
-        }
-
-        dst = &texData[gx + gy * width];
-        gw += pad * 2;
-        gh += pad * 2;
-
-        size_t bytes = gw * gh * sizeof(float) * 3;
-        if (m_sdfBuffer.size() < bytes) {
-            m_sdfBuffer.resize(bytes);
-        }
-
-        sdfBuildDistanceFieldNoAlloc(dst, width, SDF_WIDTH,
-                                     dst, gw, gh, width,
-                                     &m_sdfBuffer[0]);
-
-        texture.setDirty(gy, gh);
-    }
-
-    void releaseAtlas(std::bitset<max_textures> _refs) {
-        if (!_refs.any()) { return; }
-        std::lock_guard<std::mutex> lock(m_mutex);
-        for (size_t i = 0; i < m_batches.size(); i++) {
-            if (!_refs[i]) { continue; }
-
-            if (--m_atlasRefCount[i] == 0) {
-                LOG("CLEAR ATLAS %d", i);
-                m_atlas.clear(i);
-                m_batches[i].texData.assign(texture_size * texture_size, 0);
-            }
-        }
-    }
-    void lockAtlas(std::bitset<max_textures> _refs) {
-        if (!_refs.any()) { return; }
-        std::lock_guard<std::mutex> lock(m_mutex);
-        for (size_t i = 0; i < m_batches.size(); i++) {
-            if (_refs[i]) { m_atlasRefCount[i]++; }
-        }
-    }
-
-    alf::FontManager m_alfons;
-    alf::GlyphAtlas m_atlas;
-
-    std::shared_ptr<alf::Font> m_font;
-    std::vector<unsigned char> m_sdfBuffer;
-    std::vector<GlyphBatch> m_batches;
-    std::shared_ptr<VertexLayout> m_vertexLayout;
-    std::array<int, max_textures> m_atlasRefCount = {{0}};
-
-    std::mutex m_mutex;
-};
-
-// Not actually used as VboMesh!
-// Just keeps labels and vertices for all Labels of a tile
-struct LabelContainer : public LabelSet, public StyledMesh {
-
-    LabelContainer(AlfonsContext& _ctx) : context(_ctx) {}
-
-    ~LabelContainer() override {
-        context.releaseAtlas(atlasRefs);
-    }
-
-    AlfonsContext& context;
-    std::vector<GlyphQuad> quads;
-    std::bitset<max_textures> atlasRefs;
-
-    void draw(ShaderProgram& _shader) override {}
-    size_t bufferSize() override { return 0; }
-
-    void setLabels(std::vector<std::unique_ptr<Label>>& _labels,
-                   std::vector<GlyphQuad>& _quads) {
-
-        typedef std::vector<std::unique_ptr<Label>>::iterator iter_t;
-
-        m_labels.reserve(_labels.size());
-        m_labels.insert(m_labels.begin(),
-                        std::move_iterator<iter_t>(_labels.begin()),
-                        std::move_iterator<iter_t>(_labels.end()));
-
-        quads.reserve(_quads.size());
-        quads.insert(quads.begin(),
-                     _quads.begin(),
-                     _quads.end());
-
-        for (auto& q : quads) {
-            atlasRefs.set(q.atlas);
-        }
-
-        context.lockAtlas(atlasRefs);
-    }
-};
-
-
-TextStyle::TextStyle(std::string _name, bool _sdf, Blending _blendMode, GLenum _drawMode) :
-    Style(_name, _blendMode, _drawMode), m_sdf(_sdf),
-    m_context(std::make_shared<AlfonsContext>()) {}
-
-TextStyle::~TextStyle() {}
-
-void TextStyle::constructVertexLayout() {
-    m_vertexLayout = vertexLayout();
+    m_context->setVertexLayout(m_vertexLayout);
 }
 
 void TextStyle::constructShaderProgram() {
@@ -344,13 +57,13 @@ void TextStyle::constructShaderProgram() {
 }
 
 void TextStyle::onBeginDrawFrame(const View& _view, Scene& _scene, int _textureUnit) {
-    if (m_context->m_batches.empty()) { return; }
+    if (m_context->batches().empty()) { return; }
 
     {
         std::lock_guard<std::mutex> lock(m_context->m_mutex);
         int tex = 0;
 
-        for (auto& batch : m_context->m_batches) {
+        for (auto& batch : m_context->batches()) {
             bool dirty = false;
 
             if (batch.dirty || !batch.texture.isValid()) {
@@ -365,14 +78,14 @@ void TextStyle::onBeginDrawFrame(const View& _view, Scene& _scene, int _textureU
 
             LOG(">>> refcount:%d texture:%d/%d texid:%d upload:%d - buffersize:%d",
                 m_context->m_atlasRefCount[tex], tex,
-                m_context->m_batches.size(),
+                m_context->batches().size(),
                 batch.texture.getGlHandle(),
                 dirty, batch.mesh->bufferSize());
             tex++;
         }
     }
     m_shaderProgram->setUniformf("u_uv_scale_factor",
-                                 glm::vec2(1.0f / texture_size));
+                                 glm::vec2(1.0f / textureSize));
     m_shaderProgram->setUniformi("u_tex", 0);
     m_shaderProgram->setUniformMatrix4f("u_ortho", _view.getOrthoViewportMatrix());
 
@@ -380,13 +93,13 @@ void TextStyle::onBeginDrawFrame(const View& _view, Scene& _scene, int _textureU
 }
 
 void TextStyle::onEndDrawFrame() {
-    if (m_context->m_batches.empty()) { return; }
+    if (m_context->batches().empty()) { return; }
 
     std::lock_guard<std::mutex> lock(m_context->m_mutex);
 
     if (m_sdf) {
         m_shaderProgram->setUniformi("u_pass", 1);
-        for (auto& gt : m_context->m_batches) {
+        for (auto& gt : m_context->batches()) {
             if (gt.mesh->ready()) {
                 gt.texture.bind(0);
                 gt.mesh->draw(*m_shaderProgram);
@@ -395,7 +108,7 @@ void TextStyle::onEndDrawFrame() {
         m_shaderProgram->setUniformi("u_pass", 0);
     }
 
-    for (auto& gt : m_context->m_batches) {
+    for (auto& gt : m_context->batches()) {
         if (gt.mesh->ready()) {
 
             gt.texture.bind(0);
@@ -733,99 +446,6 @@ TextStyle::Parameters TextStyleBuilder::applyRule(const DrawRule& _rule,
 
 std::unique_ptr<StyleBuilder> TextStyle::createBuilder() const {
     return std::make_unique<TextStyleBuilder>(*this);
-}
-
-
-using namespace LabelProperty;
-
-TextLabel::TextLabel(Label::Transform _transform, Type _type, glm::vec2 _dim,
-                     LabelContainer& _labelContainer, Range _vertexRange,
-                     Label::Options _options, FontMetrics _metrics,
-                     int _nLines, Anchor _anchor, glm::vec2 _quadsLocalOrigin)
-    : Label(_transform, _dim, _type, _vertexRange, _options),
-      m_metrics(_metrics),
-      m_nLines(_nLines),
-      m_labelContainer(_labelContainer),
-      m_quadLocalOrigin(_quadsLocalOrigin) {
-
-    if (m_type == Type::point) {
-        glm::vec2 halfDim = m_dim * 0.5f;
-        switch(_anchor) {
-            case Anchor::left: m_anchor.x -= halfDim.x; break;
-            case Anchor::right: m_anchor.x += halfDim.x; break;
-            case Anchor::top: m_anchor.y -= halfDim.y; break;
-            case Anchor::bottom: m_anchor.y += halfDim.y; break;
-            case Anchor::bottom_left: m_anchor += glm::vec2(-halfDim.x, halfDim.y); break;
-            case Anchor::bottom_right: m_anchor += halfDim; break;
-            case Anchor::top_left: m_anchor -= halfDim; break;
-            case Anchor::top_right: m_anchor += glm::vec2(halfDim.x, -halfDim.y); break;
-            case Anchor::center: break;
-        }
-    }
-}
-
-void TextLabel::updateBBoxes(float _zoomFract) {
-    glm::vec2 obbCenter;
-
-    if (m_type == Type::line) {
-        m_xAxis = glm::vec2(cos(m_transform.state.rotation), sin(m_transform.state.rotation));
-        m_yAxis = glm::vec2(-m_xAxis.y, m_xAxis.x);
-    }
-
-    obbCenter = m_transform.state.screenPos;
-    // move forward on line by half the text length
-    obbCenter += m_xAxis * m_dim.x * 0.5f;
-    // move down on the perpendicular to estimated font baseline
-    obbCenter += m_yAxis * m_dim.y * 0.5f;
-    // ajdust with local origin of the quads
-    obbCenter += m_yAxis * m_quadLocalOrigin.y;
-    obbCenter += m_xAxis * m_quadLocalOrigin.x;
-
-    m_obb = OBB(obbCenter.x, obbCenter.y, m_transform.state.rotation,
-                m_dim.x + m_options.buffer, m_dim.y + m_options.buffer);
-    m_aabb = m_obb.getExtent();
-}
-
-void TextLabel::align(glm::vec2& _screenPosition, const glm::vec2& _ap1, const glm::vec2& _ap2) {
-
-    switch (m_type) {
-        case Type::debug:
-        case Type::point:
-            // modify position set by updateScreenTransform()
-            _screenPosition.x -= m_dim.x * 0.5f + m_quadLocalOrigin.x;
-            _screenPosition.y -= m_metrics.descender;
-
-            if (m_nLines > 1) {
-                _screenPosition.y -= m_dim.y * 0.5f;
-                _screenPosition.y += (m_metrics.lineHeight + m_metrics.descender);
-            }
-            _screenPosition += m_anchor;
-            break;
-        case Type::line: {
-            // anchor at line center
-            _screenPosition = (_ap1 + _ap2) * 0.5f;
-
-            // move back by half the length (so that text will be drawn centered)
-            glm::vec2 direction = glm::normalize(_ap1 - _ap2);
-            _screenPosition += direction * m_dim.x * 0.5f;
-            _screenPosition += m_yAxis * (m_dim.y * 0.5f + m_metrics.descender);
-            break;
-        }
-    }
-}
-
-void TextLabel::pushTransform() {
-    if (!visibleState()) { return; }
-
-    auto& batches = m_labelContainer.context.m_batches;
-    auto state = m_transform.state.vertex();
-
-    auto it = m_labelContainer.quads.begin() + m_vertexRange.start;
-    auto end = it + m_vertexRange.length;
-
-    for (; it != end; ++it) {
-        batches[it->atlas].mesh->pushQuad(*it, state);
-    }
 }
 
 }
